refactor: move classrules
