+++ dev/com.ibm.ws.jdbc_fat_krb5/fat/src/com/ibm/ws/jdbc/fat/krb5/FATSuite.java
@@ -1,10 +1,10 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2021 IBM Corporation and others.
+ * Copyright (c) 2020, 2025 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -14,6 +14,7 @@ package com.ibm.ws.jdbc.fat.krb5;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
+import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
@@ -21,7 +22,7 @@ import org.testcontainers.containers.Network;
 
 import com.ibm.websphere.simplicity.log.Log;
 import com.ibm.ws.jdbc.fat.krb5.containers.KerberosContainer;
-import com.ibm.ws.jdbc.fat.krb5.containers.KerberosPlatformRule;
+import com.ibm.ws.jdbc.fat.krb5.rules.KerberosPlatformRule;
 
 import componenttest.containers.TestContainerSuite;
 import componenttest.custom.junit.runner.AlwaysPassesTest;
@@ -39,18 +40,11 @@ public class FATSuite extends TestContainerSuite {
     public static Network network;
     public static KerberosContainer krb5;
 
-    static {
-        // Needed for IBM JDK 8 support.
-        java.lang.System.setProperty("com.ibm.jsse2.overrideDefaultTLS", "true");
-    }
+    @ClassRule
+    public static KerberosPlatformRule krbRule = new KerberosPlatformRule();
 
     @BeforeClass
     public static void startKerberos() throws Exception {
-        if (!KerberosPlatformRule.shouldRun(null)) {
-            // bucket will not run any tests, skip
-            return;
-        }
-
         network = Network.newNetwork();
         krb5 = new KerberosContainer(network);
         krb5.start();
@@ -58,11 +52,6 @@ public class FATSuite extends TestContainerSuite {
 
     @AfterClass
     public static void tearDown() throws Exception {
-        if (!KerberosPlatformRule.shouldRun(null)) {
-            // bucket will not run any tests, skip
-            return;
-        }
-
         Exception firstError = null;
 
         try {
