+++ dev/com.ibm.ws.jdbc_fat_krb5/fat/src/com/ibm/ws/jdbc/fat/krb5/OracleKerberosTest.java
@@ -22,16 +22,13 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
-import org.junit.rules.TestRule;
-import org.junit.runner.Description;
 import org.junit.runner.RunWith;
-import org.junit.runners.model.Statement;
 
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
 import com.ibm.websphere.simplicity.log.Log;
-import com.ibm.ws.jdbc.fat.krb5.containers.KerberosPlatformRule;
 import com.ibm.ws.jdbc.fat.krb5.containers.OracleKerberosContainer;
+import com.ibm.ws.jdbc.fat.krb5.rules.IBMJava8Rule;
 
 import componenttest.annotation.AllowedFFDC;
 import componenttest.annotation.MaximumJavaLevel;
@@ -41,7 +38,6 @@ import componenttest.custom.junit.runner.FATRunner;
 import componenttest.custom.junit.runner.Mode;
 import componenttest.custom.junit.runner.Mode.TestMode;
 import componenttest.topology.impl.JavaInfo;
-import componenttest.topology.impl.JavaInfo.Vendor;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
 import jdbc.krb5.oracle.web.OracleKerberosTestServlet;
@@ -62,9 +58,6 @@ public class OracleKerberosTest extends FATServletClient {
     @Server("com.ibm.ws.jdbc.fat.krb5.oracle")
     @TestServlet(servlet = OracleKerberosTestServlet.class, contextRoot = APP_NAME)
     public static LibertyServer server;
-
-    @ClassRule
-    public static KerberosPlatformRule skipRule = new KerberosPlatformRule();
 
     @ClassRule
     public static IBMJava8Rule skipOnIBMJava8 = new IBMJava8Rule();
@@ -152,35 +145,6 @@ public class OracleKerberosTest extends FATServletClient {
             config.getKerberos().keytab = originalKeytab;
             updateConfigAndWait(config);
         }
-    }
-
-    private static class IBMJava8Rule implements TestRule {
-
-        @Override
-        public Statement apply(Statement stmt, Description desc) {
-            return new Statement() {
-                @Override
-                public void evaluate() throws Throwable {
-                    if (shouldRun(desc)) {
-                        stmt.evaluate();
-                    }
-                }
-            };
-        }
-
-        public static boolean shouldRun(Description desc) {
-            Class<?> c = desc == null ? KerberosPlatformRule.class : desc.getTestClass();
-            String m = (desc == null || desc.getMethodName() == null) ? "shouldRun" : desc.getMethodName();
-
-            JavaInfo java = JavaInfo.forCurrentVM();
-            if (java.majorVersion() == 8 && java.vendor() == Vendor.IBM) {
-                Log.info(c, m, "Skipping tests because Oracle JDBC driver does not work with IBM JDK 8");
-                return false;
-            }
-
-            return true;
-        }
-
     }
 
     private void updateConfigAndWait(ServerConfiguration config) throws Exception {
